pyproject.toml
[project]
name = "coc-discord-bot"
version = "0.1.0"
description = "A Discord bot that acts as a TRPG Game Master, powered by GitHub Copilot SDK"
requires-python = ">=3.10"

dependencies = [
    "discord.py>=2.4",
    "github-copilot-sdk",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "xai-sdk",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
trpg-bot = "bot.main:main"
